Migrate Java WebSocketClient tests to use blocking and non-blocking contracts
