+++ http4k-client/websocket/build.gradle.kts
@@ -6,4 +6,5 @@ dependencies {
     api("org.java-websocket:Java-WebSocket:_")
     testImplementation(project(path = ":http4k-core", configuration ="testArtifacts"))
     testImplementation(project(path = ":http4k-server-jetty"))
+    testImplementation(project(path = ":http4k-realtime-core", configuration ="testArtifacts"))
 }
