+++ http4k-client/websocket/src/test/kotlin/org/http4k/client/WebsocketClientNonBlockingTest.kt
@@ -1,0 +1,11 @@
+package org.http4k.client
+
+import org.http4k.server.Jetty
+import org.http4k.websocket.NonBlockingWebsocketClientContract
+
+class WebsocketClientNonBlockingTest : NonBlockingWebsocketClientContract(
+    serverConfig = Jetty(0),
+    websocketFactory = { uri, headers, onError, onConnect ->
+        WebsocketClient.nonBlocking(uri, headers, onError = onError, onConnect = onConnect)
+    }
+)
